deploy: same deployment as for csi-driver-host-path
Using a script makes it possible to configure images without manually
editing .yaml files. This is necessary for running in a CI and more
user-friendly.

The RBAC file doesn't need to be copied into the repo when using that
script.

The script itself is a simplified version of the corresponding script
in the hostpath repo:
https://github.com/kubernetes-csi/csi-driver-host-path/blob/a72d13d1dd2d6ed87ff297790514340df97a6a35/deploy/util/deploy-hostpath.sh