Update csi-nfs-controller.yaml
Host network is needed for controller as the controller pod initial tries to mount it to nfs to create a directory structure. And for this it wont be able to access the rpc.statd on the underlying host.  

Mounting command: mount
Mounting arguments: -t nfs -o hard,nfsvers=3 192.168.125.3:/FS-SDC-U-TB4-01 /tmp/pvc-aacf7ffe-922b-4aae-aed6-2828e0234d87
Output: mount.nfs: rpc.statd is not running but is required for remote locking.
mount.nfs: Either use '-o nolock' to keep locks local, or start statd.