prow.sh: use KinD main for latest Kubernetes
KinD 0.10.0 fails to bring up a Kubernetes cluster for the latest (>
1.21.0-beta1) sources. Such breakage is expected and will be dealt
with on the main branch of KinD. Therefore we should always use when
building a cluster from source code. We have been lucky in the past
that the released versions of KinD also happend to support the latest
Kubernetes.

The downside is that temporary regressions in KinD will also affect
the Prow jobs that use the latest Kubernetes, but this is not new and
could also have happened because of regressions in Kubernetes.

While at it, building KinD gets fixed to use the configured Go
version. It happened to use Go from the default PATH.

The most recent KinD no longer supports the --type parameter, so we
have to filter that out when using KinD main.