prow.sh: add CSI_PROW_DEPLOYMENT_SUFFIX
This can be used in Prow jobs which want to let prow.sh find the
deployment directory based on the Kubernetes version but don't want
the default "kubernetes-x.yy" deployments.

The alternative would be to let Prow jobs set CSI_PROW_DEPLOYMENT, but
then the fallback to kubernetes-latest wouldn't work.