prow.sh: more flexible CSI_PROW_DEPLOYMENT
CSI_PROW_DEPLOYMENT was set explicitly for a variety of jobs. When
testing an upcoming Kubernetes release, those jobs failed because an
explicit CSI_PROW_DEPLOYMENT=kubernetes-1.21 was not matched to
kubernetes-latest and thus failed because such a deployment usually
only gets added much later.

This must have been broken for a long time.