prow.sh: fix mock testing
Mock testing was added in 74cfbc97 with the goal of increasing
test coverage and ensuring that the sidecars and driver work once
we try to use them in Kubernetes.

It turned out that this is problematic in "canary-on-<some old
Kubernetes>" jobs because those old Kubernetes version sometimes have
incompatible deployments (-provisioner argument in
external-provisioner!) because they did mock testing with much older
sidecars. Only testing against Kubernetes master is (relatively) save
and also more useful because that is where intend to use the sidecars.

Another issue was that filtering of JUnit entries did not preserve the
new mock volume entries.