build.make: support binaries outside of cmd, with optional go.mod
Setting CMDS_DIR in the top-level Makefile before including build.make will
ensure that binaries are picked up from that other directory. This is useful
for lib-volume-populator where an example populator is provided under
"example".

The commands are no longer expected to share the go.mod file with the
project. Entering the command directory and using `go build .` will pick up a
local go.mod if it exists and otherwise fall back to the project's go.mod.